Move abstract from @chalinlink/gauntlet-terra-contracts into @chainlink/gauntlet-terra
Still in process of migrating/refactoring dependencies
