+++ packages-ts/gauntlet-terra-contracts/src/commands/contracts/access_controller/removeAccess.ts
@@ -1,8 +1,12 @@
 import { BN } from '@chainlink/gauntlet-core/dist/utils'
 import { AccAddress } from '@terra-money/terra.js'
-import { AbstractInstruction, instructionToCommand } from '../../abstract/executionWrapper'
+import { abstract } from '../..'
+import {
+  AbstractInstruction,
+  instructionToCommand,
+} from '@chainlink/gauntlet-terra/dist/commands/abstract/executionWrapper'
 import { CATEGORIES } from '../../../lib/constants'
-import { CONTRACT_LIST } from '../../../lib/contracts'
+import { CONTRACT_LIST, getContract } from '../../../lib/contracts'
 
 type CommandInput = {
   address: string
